+++ web/app/router.jsx
@@ -2,5 +2,5 @@ var React = require('react/addons');
 var Router = require('react-router');
 
 Router.run(require('./routes.jsx'), Router.HistoryLocation, function routerRun(Handler, routerState) {
-  React.render(<Handler routerState={routerState} />, document.body);
+  React.render(<Handler routerState={routerState} />, document.querySelector('.react-bootstrap-element'));
 });